changes in the kafka mashling
1. added the FLOGO_EMBED env var to false value. this will create the
flogo.json when mashling binary is created. (this change is intended to
be temporary since the mashling app, by default, will not create the
separate flogo.json. Instead it should be embedded inside the binary.)

2. take a env variable for KAFKA broker URL. it was hardcoded earlier
and it would not work in different environments.

3. include instruction in README to set the KAFKA broker URL before
creating the kafka mashling.

3. include instruction to check container logs from the kafka container
to see ingress of messages on the consumer.